Ticket #719: Add a priority system to the fader module system.
Fader modules can now return an optional priority parameter as the 2nd
return from their GetOpacity function.  If no priority is returned then
it will be treated as priority 0.  The lowest opacity of the highest
priority module will be used.  This allows modules to be written to
force the opacity to be higher instead of simply lower.