Fix Ticket #726: PetHeaders being used on non-pet based groups.
If you copy a group that was setup to use the pet header (Leave gaps for
pets) then this setting would be copied and applied to non-pet groups.
So if you copied a pet group and tried to make a raid then only raid
members with pets would be shown.