Ticket #631: Allow Health and PowerBars to be animated/faded over time.
This is similar to the BarFader module in PitBull3.  However, in this
case it has been added to the BetterStatusBar implementation.  Bar
modules that want to allow their bars to be animated should set
allow_animations on their module object.

This is also implemented using the animation system added in WoW 3.1.x.
So it should be reasonably efficient.  Bars that are not set to animate
and/or fade incur no overhead.