Fix problems canceling buffs from the Blizzard BuffFrame.
In particular, we were taining the BuffFrame by tainting the PlayerFrame
due to touching PlayerFrame.unit and due to the way our HideBlizzard
module dealt with it.  The unfortunate downside of this is we no
longer turn off events for the Blizzard PlayerFrame so it is updating
even though we don't use it. D: