HideBlizzard: Tweak raid hide/show feature.  Defaults to off now.
Rational for changing the default is as follows.

A lot of people do not use PitBull4 for raid frames.  If they're using
some other addon it should hide this, not us.  Same reason we currently
don't have a Raid group already setup.

Additionally, many people were confused by not being able to find the
Raid Manger.  Since the Raid Manager lets you do set flares and do role
checks which are not provided anywhere else or in PitBull4, we also
should not be hiding by default.

People who want to hide only the raid frames and not the manager should
use the Hide button on the Raid Manager.  If you want both gone then you
can use the setting under Modules, Hide Blizzard.

This commit also fixes our hiding of the Raid Manager/Frames to
no longer break the Raid Manager.  It will properly show all the
features if you are a raid officer if you show it after having it
hidden.

Finally, this commit makes it so when we do not cause the Raid Manager
Hide/Show buttons to be out of sync with the actual state of the Raid
Frames.