Full fix to HideBlizzard problems with CompactPartyFrame in 4.1
Note: The option to hide Party frames will be disabled if you have
"Use Raid-Style Party Frames" checked in Blizzard Interface Options
under Unit Frames.  You can control these frames with the Raid checkbox.
If you want to keep the Raid Manager/Party Members, keep this unchecked
and use Hide/Show button inside the Manager.

I know this is a step back in features but Blizzard has reparented the
CompactPartyFrame to this RaidManager and it's a pain to reparent
something like this to avoid that.