Fix Bug #864 - Group Config Mode Doesn't Work.
Major refactoring of the way config mode works.  It should be faster,
though it will still take some time if you go straight into raid40 mode
without any of the frames existing.  However, subsequent changes in
config mode states after that should be pretty quick.  One side
effect of this is that the farmes won't immediately have the fake
units assigned but will pick them up very shortly after.

(Note to self: This commit differs from the anchor branch commit by the
removal of changes to the new/remove config options for units and by the
removal of logic in SwapGroupTemplate to check that the group is a pet
group)
Conflicts:
	GroupHeader.lua
	Options/Units.lua