Fix flickering Portraits on some frames.
The original fix that I'd made for the issue that happened after a scale
or shape change was wrong.  The real problem was in our anchor code.  So
I've reverted that change here.  That hack was ultimately causing the
flashing since the effective scale code was triggering on some frames
everytime.

The real problem behind the scale/shape change was that the layout code
assumed that GetHeight/Width returned the value passed to
SetHeight/Width after a ClearAllPoints, which it does not if the
previous anchors changed the size of the frame.  The fix for this is to
set a temporary bogus anchor to allow access to these values.  I'm not
100% thrilled with the solution to this, but I think the alternatives
get even more ugly.  The layout code does run in combat, so I'll likely
review this change again.