Fix Ticket #600: Performance issue with pirate day easter egg.
I just disabled the dress up model usage in combat.  It seems to avoid
the problem with lag due to the use of that model since the lag is
caused by the creation of new models.  If you're raiding or doing a
dungeon your group is still going to show as pirates since their
portrait will get drawn out of combat generally.